backend/requirements.txt
anyio==3.6.1
bcrypt==4.0.0
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
dnspython==2.2.1
ecdsa==0.18.0
fastapi==0.83.0
h11==0.13.0
idna==3.4
pytest==7.1.3
passlib==1.7.4
pyasn1==0.4.8
pycodestyle==2.9.1
pydantic==1.10.2
pymongo==4.2.0
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
requests==2.28.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.19.1
toml==0.10.2
typing_extensions==4.3.0
urllib3==1.26.12
uvicorn==0.18.3
